Pull summary into the ProjectRegistration screen
